requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bleach==5.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.4
django-advanced-filters==2.0.0
django-autoslug==1.9.9
django-bleach==3.1.0
django-ckeditor==6.7.1
django-countries==7.6.1
django-crispy-forms==2.1
django-filter==24.2
django-js-asset==2.2.0
django-stdimage==6.0.2
djangorestframework==3.15.1
idna==3.7
Markdown==3.5.2
pillow==10.3.0
requests==2.31.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tinycss2==1.1.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
